pyproject.toml
[tool.poetry]
name = "ferramentasspy"
version = "0.1.0"
description = ""
authors = ["Juliano Amaral Gouveia <julianoamg.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-model-utils = "^4.3.1"
django-bootstrap-v5 = "^1.0.11"
psycopg2-binary = "^2.9.6"
user-agents = "^2.2.0"
django-axes = "^5.41.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
